+++ assets/js/app.js
@@ -1,14 +1,3 @@
-// We need to import the CSS so that webpack will load it.
-// The MiniCssExtractPlugin is used to separate it out into
-// its own CSS file.
-import "../css/app.scss"
-
-// webpack automatically bundles all modules in your
-// entry points. Those entry points can be configured
-// in "webpack.config.js".
-//
-// Import deps with the dep name or local files with a relative path, for example:
-//
 import { Socket } from "phoenix"
 import socket from "./socket"
 import "phoenix_html"
@@ -18,8 +7,8 @@ import Hooks from "./hooks"
 
 let csrfToken = document.querySelector("meta[name='csrf-token']").getAttribute("content")
 
-import {Interpolate} from "./interpolate"
-import {ColorMap} from "./colormap"
+import Interpolate from "./interpolate"
+import ColorMap from "./colormap"
 
 window.Interpolate = Interpolate;
 window.ColorMap = ColorMap;