+++ lib/open890_web/endpoint.ex
@@ -24,7 +24,7 @@ defmodule Open890Web.Endpoint do
     at: "/",
     from: :open890,
     gzip: false,
-    only: ~w(css fonts images js favicon.ico robots.txt)
+    only: ~w(assets css fonts images js favicon.ico robots.txt)
 
   # Code reloading can be explicitly enabled under the
   # :code_reloader configuration of your endpoint.